arm64: Widen hwcap to be 64 bit
Under arm64 elf_hwcap is a 32 bit quantity, but it is stored in
a 64 bit auxiliary ELF field and glibc reads hwcap as 64 bit.

This patch widens elf_hwcap to be 64 bit.

Git-commit: 25804e6a96681d5d2142058948e218999e4f547c
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git

+++ arch/arm64/kernel/setup.c
@@ -59,7 +59,7 @@
 unsigned int processor_id;
 EXPORT_SYMBOL(processor_id);
 
-unsigned int elf_hwcap __read_mostly;
+unsigned long elf_hwcap __read_mostly;
 EXPORT_SYMBOL_GPL(elf_hwcap);
 
 unsigned int boot_reason;
